requirements.txt
databases==0.6.0
discord.py==1.7.3
pillow==9.2.0
requests==2.28.1
asyncpg==0.26.0
